repair borked tests
