+++ api/src/opentrons/protocol_api/types.py
@@ -1,5 +1,4 @@
-from typing import Any, Callable, Dict, List
-from dataclasses import dataclass
+from typing import Any, Callable, Dict
 from .labware import Labware
 from .contexts import InstrumentContext, ProtocolContext, \
     MagneticModuleContext, TemperatureModuleContext
@@ -17,17 +16,3 @@ PipetteHandler = Callable[[Instruments, LoadedLabware, Any], None]
 MagneticModuleHandler = Callable[[MagneticModuleContext, Any], None]
 
 TemperatureModuleHandler = Callable[[TemperatureModuleContext, Any], None]
-
-# Deck definition types
-
-
-@dataclass
-class CalibrationPosition:
-    """
-    A point on the deck of a robot that is used to calibrate
-    aspects of the robot's movement system as defined by
-    opentrons/shared-data/deck/schemas/2.json
-    """
-    id: str
-    position: List[float]
-    displayName: str
